+++ app/Http/Controllers/MessagesController.php
@@ -10,6 +10,7 @@ class MessagesController extends Controller
 {
     public function index()
     {
+        $this->authorize('interactAsMaster'); // If false, it'll display 403
         $messages = Messages::all();
 
         return view('dashboard.admin.message.index', compact('messages'));
@@ -17,15 +18,18 @@ class MessagesController extends Controller
 
     public function show(Messages $message)
     {
+        $this->authorize('interactAsMaster'); // If false, it'll display 403
         return view('dashboard.admin.message.show', compact('message'));
     }
 
     public function create() {
+        $this->authorize('interactAsMaster'); // If false, it'll display 403
         return view('pages.contact');
     }
 
     public function update(Request $request, Messages $message)
     {
+        $this->authorize('interactAsMaster'); // If false, it'll display 403
         $message->update($request->validate([
             'status' => [
                 'required',
@@ -56,6 +60,7 @@ class MessagesController extends Controller
 
     public function destroy(Messages $message)
     {
+        $this->authorize('interactAsMaster'); // If false, it'll display 403
         $message->delete();
 
         return redirect()->route('message.index');
